Regenerar clientes

+++ gpaserveisWeb/src/main/java/es/bcn/gpa/gpaserveis/business/impl/ExpedientsServiceImpl.java
@@ -212,7 +212,7 @@ public class ExpedientsServiceImpl implements ExpedientsService {
 		}
 
 		try {
-			SollicitudsRDTO sollicitudsRDTO = sollicitudsApi.consultarDadesSollicituds(id.intValue());
+			SollicitudsRDTO sollicitudsRDTO = sollicitudsApi.consultarDadesSollicituds(id);
 
 			if (log.isDebugEnabled()) {
 				log.debug("consultarDadesSollicitud(BigDecimal) - fi"); //$NON-NLS-1$
